\name{coef.summary.tssdr}
\alias{coef.summary.tssdr}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Coefficients of a summary.tssdr object
}
\description{
Extracts the estimated signal separation matrix from an object of class summary.tssdr.
}
\usage{
\method{coef}{summary.tssdr}(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{Object of class summary.tssdr.}
  \item{\dots}{Further arguments to be passed to or from methods.}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\author{Markus Matilainen}
\seealso{
\code{\link[tsBSS]{summary.tssdr}}
}
\keyword{methods}
