\name{components.summary.tssdr}
\alias{components.summary.tssdr}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Directions of an object of class summary.tssdr}
\description{
Gives directions of an object of class summary.tssdr.
}
\usage{
\method{components}{summary.tssdr}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{Object of class summary.tssdr.}
  \item{\dots}{Further arguments to be passed to or from methods.}
}
\author{Markus Matilainen}
\seealso{
\code{\link[tsBSS]{summary.tssdr}}
}
\keyword{methods}
