\name{components.tssdr}
\alias{components.tssdr}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Directions of an object of class tssdr}
\description{
Gives directions of an object of class tssdr.
}
\usage{
\method{components}{tssdr}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{Object of class tssdr.}
  \item{\dots}{Further arguments to be passed to or from methods.}
}
\author{Markus Matilainen}
\seealso{
\code{\link[tsBSS]{tssdr}}
}
\keyword{methods}
