\name{print.bssvol}
\alias{print.bssvol}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Printing an object of class bssvol}
\description{
Prints an object of class bssvol. It prints the unmixing matrix W and the used lags k.
}
\usage{
\method{print}{bssvol}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{An object of class bssvol.}
  \item{\dots}{Further arguments to be passed to or from methods.}
}
\author{Markus Matilainen}
\seealso{
\code{\link[JADE]{print.bss}}
}
\keyword{methods}
