\name{print.lbtest}
\alias{print.lbtest}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Printing an object of class lbtest}
\description{
Prints an object of class lbtest.
}
\usage{
\method{print}{lbtest}(x, digits = 3, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{An object of class lbtest.}
  \item{digits}{The chosen number of digits.}
  \item{\dots}{Further arguments to be passed to or from methods.}
}
\author{Markus Matilainen}
\seealso{
\code{\link[tsBSS]{lbtest}}
}
\keyword{methods}
