\name{print.summary.tssdr}
\alias{print.summary.tssdr}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Printing an object of class summary.tssdr}
\description{
Prints an object of class summary.tssdr. It prints all elements of the list of class summary.tssdr except the component S, which is the matrix of the chosen directions or a vector, if only one direcction is chosen.
}
\usage{
\method{print}{summary.tssdr}(x, digits = 3, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{Object of class summary.tssdr.}
  \item{digits}{Chosen number of digits.}
  \item{\dots}{Further arguments to be passed to or from methods.}
}
\author{Markus Matilainen}
\seealso{
\code{\link[tsBSS]{summary.tssdr}}
}
\keyword{methods}
